finishAuth

+++ src/main/java/com/re/cinema_manager/repository/UserRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     // Hàm này giúp Spring tự động tạo lệnh: SELECT * FROM users WHERE username = ?
     // Dùng để kiểm tra xem tài khoản đã tồn tại chưa khi đăng ký, và tìm user khi đăng nhập
@@ -15,3 +16,4 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     // Tương tự, dùng để kiểm tra xem email đã có ai dùng chưa
     boolean existsUserByUsername(String username);
 }
+
